% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model.R
\name{derive_model}
\alias{derive_model}
\title{Derive a \code{\link[=base_model]{base_model()}} from a template list.
Checks for type, mode and class of the object.}
\usage{
derive_model(template)
}
\arguments{
\item{template}{A list of key/value pairs with the value used for type derivation.}
}
\value{
Returns a \code{\link[=base_model]{base_model()}} object.
}
\description{
Derive a \code{\link[=base_model]{base_model()}} from a template list.
Checks for type, mode and class of the object.
}
\examples{
template <- list(a_character = "hello",
                 a_model = lm(mpg ~ cyl, data = mtcars),
                 a_dataframe = mtcars,
                 a_number = 12,
                 an_integer = 12L)


my_model <- derive_model(template)

# This works
my_model(
  an_integer = 12L,
  a_character = "test",
  a_dataframe = iris,
  a_model = lm(Sepal.Width ~ Petal.Width, data = iris),
  a_number = 7
)

# This throws an error
try(
  my_model(
    an_integer = 12,
    a_character = "test",
    a_dataframe = iris,
    a_model = lm(Sepal.Width ~ Petal.Width, data = iris),
    a_number = 7
  )
)

}
