% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{.coerce_hint}
\alias{.coerce_hint}
\title{Explain why a value could not be coerced}
\usage{
.coerce_hint(name, x)
}
\arguments{
\item{name}{The primitive type's name.}

\item{x}{The value that failed.}
}
\value{
A one-line hint, or \code{NULL}.
}
\description{
Explain why a value could not be coerced
}
\examples{
rdantic:::.coerce_hint("int", 1.5)
rdantic:::.coerce_hint("int", 1e10)
rdantic:::.coerce_hint("num", "7")
}
\keyword{internal}
