% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{.comparable}
\alias{.comparable}
\title{Could two values be equal without coercion?}
\usage{
.comparable(x, v)
}
\arguments{
\item{x, v}{Atomic values.}
}
\value{
\code{TRUE} when comparing them would not stringify either side.
}
\description{
Could two values be equal without coercion?
}
\examples{
rdantic:::.comparable(1, 2)
rdantic:::.comparable(1, "2")
}
\keyword{internal}
