% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{.hints_of}
\alias{.hints_of}
\title{Collect the hints from a set of problems}
\usage{
.hints_of(r)
}
\arguments{
\item{r}{A \code{typed_problems} object.}
}
\value{
A character vector, possibly empty.
}
\description{
Collect the hints from a set of problems
}
\examples{
rdantic:::.hints_of(rdantic:::.bad("", "int", "num[1] 1.5", "not a whole number"))
}
\keyword{internal}
