% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7.R
\name{.is_s7_class}
\alias{.is_s7_class}
\title{Is this an S7 class object?}
\usage{
.is_s7_class(x)
}
\arguments{
\item{x}{Any R value.}
}
\value{
\code{TRUE} for the object \code{S7::new_class()} returns.
}
\description{
Is this an S7 class object?
}
\examples{
rdantic:::.is_s7_class(int)
}
\keyword{internal}
