% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7.R
\name{.is_s7_object}
\alias{.is_s7_object}
\title{Is this an instance of an S7 class?}
\usage{
.is_s7_object(x)
}
\arguments{
\item{x}{Any R value.}
}
\value{
\code{TRUE} for an S7 instance.
}
\description{
A class object is itself an \code{S7_object}, so it has to be excluded.
}
\examples{
rdantic:::.is_s7_object(1)
}
\keyword{internal}
