% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{.list_of_keys}
\alias{.list_of_keys}
\title{A named list with a fixed key set}
\usage{
.list_of_keys(keys, extra)
}
\arguments{
\item{keys}{A named list of types.}

\item{extra}{\code{"ignore"} or \code{"forbid"}.}
}
\value{
A type.
}
\description{
A named list with a fixed key set
}
\examples{
rdantic:::.list_of_keys(list(a = int[1]), "forbid")(list(a = 1))
}
\keyword{internal}
