% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{.map_of_any}
\alias{.map_of_any}
\title{A map of any length with one element type, keyed by name}
\usage{
.map_of_any(t)
}
\arguments{
\item{t}{The element type.}
}
\value{
A type.
}
\description{
A map of any length with one element type, keyed by name
}
\examples{
rdantic:::.map_of_any(int[1])(list(a = 1, b = 2))
}
\keyword{internal}
