% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{.members_of}
\alias{.members_of}
\title{The fields a type can be subclassed or weakened through}
\usage{
.members_of(t, what)
}
\arguments{
\item{t}{A type.}

\item{what}{The calling function, for the error message.}
}
\value{
A named list of types.
}
\description{
The fields a type can be subclassed or weakened through
}
\examples{
names(rdantic:::.members_of(struct("FieldsDemo", a = int[1]), "extend()"))
}
\keyword{internal}
