% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{.refined}
\alias{.refined}
\title{Restrict a type with a predicate}
\usage{
.refined(t, expr, env)
}
\arguments{
\item{t}{A type.}

\item{expr}{An unevaluated expression in \code{.}.}

\item{env}{The environment the expression is evaluated in.}
}
\value{
A type.
}
\description{
Restrict a type with a predicate
}
\examples{
rdantic:::.refined(int, quote(. > 0), environment())(c(1, 2))
}
\keyword{internal}
