% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{.required_fields}
\alias{.required_fields}
\title{Which fields must be supplied?}
\usage{
.required_fields(fields)
}
\arguments{
\item{fields}{A named list of types.}
}
\value{
A character vector of field names.
}
\description{
A field is optional when it has a default or when its type accepts \code{NULL}.
}
\examples{
rdantic:::.required_fields(list(a = int[1], b = int[1] | NULL, c = chr \%default\% "x"))
}
\keyword{internal}
