% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{.set_field}
\alias{.set_field}
\title{Validate and set a field}
\usage{
.set_field(x, name, value)
}
\arguments{
\item{x}{An instance.}

\item{name}{A field name.}

\item{value}{The new value.}
}
\value{
A new instance.
}
\description{
Validate and set a field
}
\examples{
rdantic:::.set_field(model("SetDemo", a = int[1])(a = 1), "a", 2)
}
\keyword{internal}
