% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{.sized}
\alias{.sized}
\title{Fix the length of a type}
\usage{
.sized(t, n)
}
\arguments{
\item{t}{A type.}

\item{n}{The required length.}
}
\value{
A type.
}
\description{
Fix the length of a type
}
\examples{
rdantic:::.sized(int, 2)(c(1, 2))
}
\keyword{internal}
