% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{.spec}
\alias{.spec}
\title{Get a type's spec}
\usage{
.spec(t)
}
\arguments{
\item{t}{A type.}
}
\value{
The spec list: \code{name}, \code{validate}, \code{schema}, plus any extras.
}
\description{
Get a type's spec
}
\examples{
names(rdantic:::.spec(int))
}
\keyword{internal}
