% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{is_another_model}
\alias{is_another_model}
\title{Check another model inside a model}
\usage{
is_another_model(.model)
}
\arguments{
\item{.model}{model to check}
}
\description{
Check another model inside a model
}
