% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{is_valid}
\alias{is_valid}
\title{Does a value satisfy a type?}
\usage{
is_valid(t, x)
}
\arguments{
\item{t}{A type, or anything \code{\link[=as_type]{as_type()}} accepts.}

\item{x}{The value to test.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Does a value satisfy a type?
}
\examples{
is_valid(int[1], 3)
is_valid(int[1], 3.5)
is_valid(chr, letters)
}
