% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keys_to_camel_case}
\alias{keys_to_camel_case}
\title{Convert keys to camel case}
\usage{
keys_to_camel_case(x, .recursive = TRUE)
}
\arguments{
\item{x}{list}

\item{.recursive}{Whether to convert keys on all levels.}
}
\description{
Convert keys to camel case
}
