% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{print.type}
\alias{print.type}
\title{Print a type}
\usage{
\method{print}{type}(x, ...)
}
\arguments{
\item{x}{A type.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a type
}
\examples{
print(int[1])
print(one_of("a", "b") \%default\% "a")
}
