% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{print.typed_fn}
\alias{print.typed_fn}
\title{Print a typed function's signature}
\usage{
\method{print}{typed_fn}(x, ...)
}
\arguments{
\item{x}{A typed function.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a typed function's signature
}
\examples{
print(fn(x = int[1], y = chr[1] \%default\% "a", ~ chr[1], { paste0(y, x) }))
}
