% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.typed_instance}
\alias{print.typed_instance}
\title{Print an instance}
\usage{
\method{print}{typed_instance}(x, ...)
}
\arguments{
\item{x}{An instance.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an instance
}
\examples{
Box <- model("Box", label = chr[1], items = list_of(chr[1]))
print(Box(label = "tools", items = list("hammer", "nail")))
}
