% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.typed_model}
\alias{print.typed_model}
\title{Print a model}
\usage{
\method{print}{typed_model}(x, ...)
}
\arguments{
\item{x}{A model.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a model
}
\examples{
print(model("PrintDemo", id = int[1], tag = chr[1] \%default\% "none"))
}
