% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{print.typed_struct}
\alias{print.typed_struct}
\title{Print a struct}
\usage{
\method{print}{typed_struct}(x, ...)
}
\arguments{
\item{x}{A struct.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a struct
}
\examples{
print(struct("PrintDemo", id = int[1], tag = chr[1] \%default\% "none"))
}
