% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{type_of}
\alias{type_of}
\title{Describe the type of a value the way rdantic names types}
\usage{
type_of(x)
}
\arguments{
\item{x}{Any R value.}
}
\value{
A single string, e.g. \code{"int[3]"}, \code{"<User>"}, \code{"data.frame[2 x 3]"}.
}
\description{
Length is part of the description, because in R it is part of the type:
\code{int[1]} and \code{int[3]} are different shapes with different bugs.
}
\examples{
type_of(1:3)
type_of(1.5)
type_of("a")
type_of(Sys.Date())
type_of(data.frame(a = 1:2))
type_of(NULL)
}
