% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{typed_fn}
\alias{typed_fn}
\title{Type for any typed function}
\usage{
typed_fn(x)
}
\arguments{
\item{x}{A value to validate.}
}
\value{
A type.
}
\description{
\code{typed_fn} accepts any function created by \code{\link[=fn]{fn()}}. Use \code{\link[=fn_type]{fn_type()}} when its
declared argument and return types must match a particular signature.
}
\examples{
identity_int <- fn(x = int[1], ~ int[1], { x })
typed_fn(identity_int)
try(typed_fn(function(x) x))
}
