% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model.R
\name{validate_fn}
\alias{validate_fn}
\title{Validate function arguments}
\usage{
validate_fn(fn)
}
\arguments{
\item{fn}{A function containing the arguments to be checked.}
}
\value{
The function's \code{environment}.
}
\description{
Validate function arguments
}
\examples{
f <- function(a = is.integer, b = is.integer) {
  validate_fn(f)
  a + b
}

f(2L, 5L)

try(f(2, 5))
}
