% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{as.list.typed_instance}
\alias{as.list.typed_instance}
\title{Convert an instance to a list}
\usage{
\method{as.list}{typed_instance}(x, ...)
}
\arguments{
\item{x}{An instance.}

\item{...}{Ignored.}
}
\value{
A named list.
}
\description{
Convert an instance to a list
}
\examples{
as.list(struct("Rgb", r = int[1], g = int[1], b = int[1])(r = 1, g = 2, b = 3))
}
