% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model.R
\name{base_model}
\alias{base_model}
\title{Create a model}
\usage{
base_model(..., .validators_before = NULL, .validators_after = NULL)
}
\arguments{
\item{...}{model parameters and their type-check functions}

\item{.validators_before}{list of validators that run before types are checked}

\item{.validators_after}{list of validators that run after types are checked}
}
\value{
model function
}
\description{
Create a model
}
\examples{
validators <- list(
  b = function(x) ifelse(x == 0, 1, x)
)

my_model <- base_model(
  a = is.numeric,
  b = is.numeric,
  .validators = validators
)

f <- function(a, b) a / b

do.call(f, my_model(a = 1, b = 0)) # 1
do.call(f, list(a = 1, b = 0)) # Inf

# Use model inside function
f <- function(a, b) {
  params <- my_model(environment())
  params$a / params$b
}

f(1, 0) # 1
}
