% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{base_model}
\alias{base_model}
\title{Create a model}
\usage{
base_model(..., .validators = NULL)
}
\arguments{
\item{...}{model parameters and their type testers}

\item{.validators}{list of validators}
}
\value{
model function
}
\description{
Create a model
}
\examples{
validators <- list(
  b = function(x) ifelse(x == 0, 1, x)
)

my_model <- base_model(
  a = is.numeric,
  b = is.numeric,
  .validators = validators
)

f <- function(a, b) a / b

do.call(f, my_model(a = 1, b = 0)) # 1
do.call(f, list(a = 1, b = 0)) # Inf

# Use model inside function
f <- function(a, b) {
  params = my_model(environment())
  params$a / params$b
}

f(1, 0) # 1
}
