% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model.R
\name{base_model}
\alias{base_model}
\title{Create a model}
\usage{
base_model(..., .validators_before = NULL, .validators_after = NULL)
}
\arguments{
\item{...}{model parameters and their type-check functions}

\item{.validators_before}{list of validators that run before types are checked}

\item{.validators_after}{list of validators that run after types are checked}
}
\value{
model function
}
\description{
Create a model
}
\examples{
# Models
my_model <- base_model(
  a = is.integer,
  b = is.integer,
  txt = is.character
)

my_model(a = 1L, b = 2L, txt = "My awesome model")
try(my_model(a = 1, b = 2L, txt = "My awesome model"))

# Validate function args
f <- function(a, b) {
  validate_args(a = is.numeric, b = is.numeric)
  a + b
}

f(4, 5)
try(f(4, "5"))
}
