% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{.list_of_any}
\alias{.list_of_any}
\title{A list of any length with one element type}
\usage{
.list_of_any(t)
}
\arguments{
\item{t}{The element type.}
}
\value{
A type.
}
\description{
A list of any length with one element type
}
\examples{
rdantic:::.list_of_any(int[1])(list(1, 2))
}
\keyword{internal}
