% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{fields}
\alias{fields}
\title{Field names of a struct or an instance}
\usage{
fields(x)
}
\arguments{
\item{x}{A struct, an instance, or anything \code{\link[=as_type]{as_type()}} accepts.}
}
\value{
A character vector, in declaration order.
}
\description{
Field names of a struct or an instance
}
\examples{
Pt <- struct("Pt", x = num[1], y = num[1])
fields(Pt)
fields(Pt(x = 1, y = 2))
}
