% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model.R
\name{model_dump}
\alias{model_dump}
\title{Modify the model representation}
\usage{
model_dump(
  .obj,
  exclude = NULL,
  include = NULL,
  exclude_null = FALSE,
  exclude_na = FALSE,
  camels = FALSE
)
}
\arguments{
\item{exclude}{A set of fields to exclude from the output.}

\item{exclude_null}{Whether to drop items with the value `NULL`.}

\item{exclude_na}{Whether to drop items with the value `NA`.}

\item{camels}{Whether to convert all keys to camel case.}

\item{incluse}{A set of fields to include in the output.}
}
\value{
list
}
\description{
Modify the model representation
}
