% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{partial}
\alias{partial}
\title{Make every field of a model optional}
\usage{
partial(.parent, .name = paste0("Partial", .spec(.parent)$name))
}
\arguments{
\item{.parent}{The model to weaken.}

\item{.name}{The new model's name.}
}
\value{
A constructor of class \code{typed_model}.
}
\description{
For PATCH-shaped input, where only the supplied keys mean anything. Defaults
are dropped as well as requirements, so "not supplied" stays distinguishable
from "set to the default".
}
\examples{
Post <- model("Post", title = chr[1], body = chr[1], draft = lgl[1] \%default\% TRUE)
PostPatch <- partial(Post)
PostPatch

patch <- PostPatch(title = "New title")
patch$title
patch$draft          # NULL, not the default
}
\seealso{
\code{\link[=model]{model()}}, \code{\link[=extend]{extend()}}
}
