% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model.R
\name{validate_args}
\alias{validate_args}
\title{Validate function arguments}
\usage{
validate_args(..., .validators_before = NULL, .validators_after = NULL)
}
\arguments{
\item{...}{Model parameters and their functions used for the type checks.
Usually only \code{fields} or \code{...} are supplied.}

\item{.validators_before}{A list of validators that run before types are checked.}

\item{.validators_after}{A list of validators that run after types are checked.}
}
\value{
base model function/
}
\description{
Validate function arguments
}
\examples{
f <- function(a, b) {
  validate_args(
    a = is.integer,
    b = is.integer
  )
  a + b
}

f(2L, 5L)

try(f(2, 5))
}
