% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{validate_args}
\alias{validate_args}
\title{Validate args inside function}
\usage{
validate_args(..., .validators = NULL)
}
\arguments{
\item{...}{model parameters and their type-check functions}
}
\value{
model function
}
\description{
Validate args inside function
}
