% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{validate_args}
\alias{validate_args}
\title{Validate args inside function}
\usage{
validate_args(..., .validators_before = NULL, .validators_after = NULL)
}
\arguments{
\item{...}{model parameters and their type-check functions}

\item{.validators_before}{list of validators that run before types are checked}

\item{.validators_after}{list of validators that run after types are checked}
}
\value{
model function
}
\description{
Validate args inside function
}
