% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model.R
\name{validate_fn}
\alias{validate_fn}
\title{Validate function parameters}
\usage{
validate_fn(fn)
}
\arguments{
\item{fn}{function containing the arguments to be checked}
}
\description{
Validate function parameters
}
